fix: SpawnerController, remove Spawner and refactored SO as SpawnerZone
