--- Assets/Prefabs/Spawner.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 8c6ecfcfe04fd8a43bae02bf181d7564
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
