--- Assets/Scripts/Components/Controllers.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: b9ce5d5186f57374fb139a90eafc56a6
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
