--- Assets/Scripts/Components/Spawner.cs.meta
@@ -1,3 +1,0 @@
-fileFormatVersion: 2
-guid: a227d68597c9418dbc840e987bc096fc
-timeCreated: 1643716468
