--- Assets/Scripts/Interfaces/BaseSpawner.cs.meta
@@ -1,3 +1,0 @@
-fileFormatVersion: 2
-guid: 9df79aa18b5a40469277052448442141
-timeCreated: 1643717393
