--- Assets/Scripts/Interfaces/ISpawner.cs.meta
@@ -1,3 +1,0 @@
-fileFormatVersion: 2
-guid: d4db1a5e3d254816a0b461be09686543
-timeCreated: 1643716354
