+++ Assets/Scripts/Settings/SpawnZone.cs
@@ -2,15 +2,18 @@ using UnityEngine;
 
 namespace Settings
 { 
-    [CreateAssetMenu(fileName = "Spawner", menuName = "Spawner Settings", order = 0)]
-    public class SpawnerSettings: ScriptableObject
+    [CreateAssetMenu(fileName = "SpawnZone", menuName = "Spawn Zone", order = 0)]
+    public class SpawnZone: ScriptableObject
     {
         [SerializeField] private float spawnPercent;
+        [Space]
         [SerializeField] private Vector2 minPoint;
         [SerializeField] private Vector2 maxPoint;
-        [SerializeField] private float angleMinPoint;
+        [Space]
+        [SerializeField, Range(0.0f, 360.0f)] private float angleMinPoint;
+        [SerializeField, Range(0.0f, 360.0f)] private float angleMaxPoint;
+        [Space] 
         [SerializeField] private float forceMinPoint;
-        [SerializeField] private float angleMaxPoint;
         [SerializeField] private float forceMaxPoint;
 
         public float SpawnPercent => spawnPercent;
