+++ Assets/Settings/SpawnZoneBottomRight.asset
@@ -10,12 +10,12 @@ MonoBehaviour:
   m_Enabled: 1
   m_EditorHideFlags: 0
   m_Script: {fileID: 11500000, guid: 6803607dbe0d472aa1646d71f1a3dcdd, type: 3}
-  m_Name: SpawnerBottomRight
+  m_Name: SpawnZoneBottomRight
   m_EditorClassIdentifier: 
   spawnPercent: 30
   minPoint: {x: 0.5, y: 0}
   maxPoint: {x: 1, y: 0}
-  angleMinPoint: 0
-  forceMinPoint: 0
-  angleMaxPoint: 0
-  forceMaxPoint: 0
+  angleMinPoint: 90
+  angleMaxPoint: 105
+  forceMinPoint: 5
+  forceMaxPoint: 5
