+++ Assets/Settings/SpawnZoneRight.asset.meta
